requirements.txt
asttokens==2.4.1
awkward==2.6.5
awkward-cpp==34
awkward1==1.0.0
cachetools==5.3.3
cramjam==2.8.3
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
fsspec==2024.6.0
importlib_metadata==7.1.0
ipython==8.18.1
jedi==0.19.1
joblib==1.4.2
matplotlib-inline==0.1.7
numpy==1.26.4
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyyml==0.0.2
six==1.16.0
stack-data==0.6.3
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uproot==5.3.7
uproot3-methods==0.10.1
uproot4==4.0.0
wcwidth==0.2.13
zipp==3.19.2
